'Added random quote functionality'

+++ build.ps1
@@ -1,4 +1,12 @@
-﻿# Remove current pdf
+﻿param([string]$m = "")
+
+# If the commit message is null, exit script
+if($m -eq ""){
+    Write-Host "Must enter a commit message";
+    exit;
+}
+
+# Remove current pdf
 if(Test-Path Resume_2018.pdf){
     Remove-Item Resume_2018.pdf
 }
@@ -17,3 +25,6 @@ cd "C:\xampp\htdocs\bgreenawald.github.io\"
 # Run the blog build script
 python generate.py
 
+# Push to git
+git add .
+git commit -m "'$m'"
